@@ -2,21 +2,22 @@ name: cached_network_image
 description: Flutter library to load and cache network images.
   Can also be used with placeholder and error widgets.
 homepage: https://github.com/Baseflow/flutter_cached_network_image
-version: 2.5.0
+version: 3.0.0-nullsafety
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_cache_manager: ^2.1.0
-  octo_image: ^0.3.0
+  flutter_cache_manager: ^3.0.0-nullsafety
+  octo_image: ^1.0.0-nullsafety
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.1
-  pedantic: ^1.9.0
-  file: ^5.2.1
+  mockito: ^5.0.0
+  build_runner: ^1.10.0
+  pedantic: ^1.10.0
+  file: ^6.1.0
 
 environment:
-  flutter: ">=1.20.0 <2.0.0"
-  sdk: ">=2.10.0 <3.0.0"
+  flutter: ">=1.24.0-10.2.pre <2.0.0"
+  sdk: ">=2.12.0-133.7.beta <3.0.0"
