@@ -16,12 +16,11 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
   /// [chunkEvents] should indicate the [ImageChunkEvent]s of the first image
   /// to show.
   MultiImageStreamCompleter({
-    @required Stream<ui.Codec> codec,
-    @required double scale,
-    Stream<ImageChunkEvent> chunkEvents,
-    InformationCollector informationCollector,
-  })  : assert(codec != null),
-        _informationCollector = informationCollector,
+    required Stream<ui.Codec> codec,
+    required double scale,
+    Stream<ImageChunkEvent>? chunkEvents,
+    InformationCollector? informationCollector,
+  })  : _informationCollector = informationCollector,
         _scale = scale {
     codec.listen((event) {
       if (_timer != null) {
@@ -54,18 +53,18 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
     }
   }
 
-  ui.Codec _codec;
-  ui.Codec _nextImageCodec;
+  ui.Codec? _codec;
+  ui.Codec? _nextImageCodec;
   final double _scale;
-  final InformationCollector _informationCollector;
-  ui.FrameInfo _nextFrame;
+  final InformationCollector? _informationCollector;
+  ui.FrameInfo? _nextFrame;
   // When the current was first shown.
-  Duration _shownTimestamp;
+  Duration? _shownTimestamp;
   // The requested duration for the current frame;
-  Duration _frameDuration;
+  Duration? _frameDuration;
   // How many frames have been emitted so far.
   int _framesEmitted = 0;
-  Timer _timer;
+  Timer? _timer;
 
   // Used to guard against registering multiple _handleAppFrame callbacks for the same frame.
   bool _frameCallbackScheduled = false;
@@ -73,13 +72,12 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
   void _switchToNewCodec() {
     _framesEmitted = 0;
     _timer = null;
-    _handleCodecReady(_nextImageCodec);
+    _handleCodecReady(_nextImageCodec!);
     _nextImageCodec = null;
   }
 
   void _handleCodecReady(ui.Codec codec) {
     _codec = codec;
-    assert(_codec != null);
 
     if (hasListeners) {
       _decodeNextFrameAndSchedule();
@@ -90,22 +88,22 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
     _frameCallbackScheduled = false;
     if (!hasListeners) return;
     if (_isFirstFrame() || _hasFrameDurationPassed(timestamp)) {
-      _emitFrame(ImageInfo(image: _nextFrame.image, scale: _scale));
+      _emitFrame(ImageInfo(image: _nextFrame!.image, scale: _scale));
       _shownTimestamp = timestamp;
-      _frameDuration = _nextFrame.duration;
+      _frameDuration = _nextFrame!.duration;
       _nextFrame = null;
-      if (_framesEmitted % _codec.frameCount == 0 && _nextImageCodec != null) {
+      if (_framesEmitted % _codec!.frameCount == 0 && _nextImageCodec != null) {
         _switchToNewCodec();
       } else {
-        final completedCycles = _framesEmitted ~/ _codec.frameCount;
-        if (_codec.repetitionCount == -1 ||
-            completedCycles <= _codec.repetitionCount) {
+        final completedCycles = _framesEmitted ~/ _codec!.frameCount;
+        if (_codec!.repetitionCount == -1 ||
+            completedCycles <= _codec!.repetitionCount) {
           _decodeNextFrameAndSchedule();
         }
       }
       return;
     }
-    final delay = _frameDuration - (timestamp - _shownTimestamp);
+    final delay = _frameDuration! - (timestamp - _shownTimestamp!);
     _timer = Timer(delay * timeDilation, _scheduleAppFrame);
   }
 
@@ -114,13 +112,12 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
   }
 
   bool _hasFrameDurationPassed(Duration timestamp) {
-    assert(_shownTimestamp != null);
-    return timestamp - _shownTimestamp >= _frameDuration;
+    return timestamp - _shownTimestamp! >= _frameDuration!;
   }
 
   Future<void> _decodeNextFrameAndSchedule() async {
     try {
-      _nextFrame = await _codec.getNextFrame();
+      _nextFrame = await _codec!.getNextFrame();
     } catch (exception, stack) {
       reportError(
         context: ErrorDescription('resolving an image frame'),
@@ -131,7 +128,7 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
       );
       return;
     }
-    if (_codec.frameCount == 1) {
+    if (_codec!.frameCount == 1) {
       // ImageStreamCompleter listeners removed while waiting for next frame to
       // be decoded.
       // There's no reason to emit the frame without active listeners.
@@ -141,7 +138,7 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
 
       // This is not an animated image, just return it and don't schedule more
       // frames.
-      _emitFrame(ImageInfo(image: _nextFrame.image, scale: _scale));
+      _emitFrame(ImageInfo(image: _nextFrame!.image, scale: _scale));
       return;
     }
     _scheduleAppFrame();
@@ -152,7 +149,7 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
       return;
     }
     _frameCallbackScheduled = true;
-    SchedulerBinding.instance.scheduleFrameCallback(_handleAppFrame);
+    SchedulerBinding.instance?.scheduleFrameCallback(_handleAppFrame);
   }
 
   void _emitFrame(ImageInfo imageInfo) {
