@@ -13,7 +13,6 @@ class FakeCacheManager extends Mock implements CacheManager {
   void throwsNotFound(String url) {
     when(getFileStream(
       url,
-      withProgress: anyNamed('withProgress'),
       headers: anyNamed('headers'),
       key: anyNamed('key'),
     )).thenThrow(HttpExceptionWithStatus(404, 'Invalid statusCode: 404',
@@ -23,7 +22,7 @@ class FakeCacheManager extends Mock implements CacheManager {
   ExpectedData returns(
     String url,
     List<int> imageData, {
-    Duration delayBetweenChunks,
+    Duration? delayBetweenChunks,
   }) {
     const chunkSize = 8;
     final chunks = <Uint8List>[
@@ -33,10 +32,10 @@ class FakeCacheManager extends Mock implements CacheManager {
 
     when(getFileStream(
       url,
-      withProgress: anyNamed('withProgress'),
+      withProgress: anyNamed('withProgress') ?? false,
       headers: anyNamed('headers'),
       key: anyNamed('key'),
-    )).thenAnswer((realInvocation) => _createResultStream(
+    )).thenAnswer((realInvocation) => createResultStream(
           url,
           chunks,
           imageData,
@@ -50,11 +49,11 @@ class FakeCacheManager extends Mock implements CacheManager {
     );
   }
 
-  Stream<FileResponse> _createResultStream(
+  Stream<FileResponse> createResultStream(
     String url,
     List<Uint8List> chunks,
     List<int> imageData,
-    Duration delayBetweenChunks,
+    Duration? delayBetweenChunks,
   ) async* {
     var totalSize = imageData.length;
     var downloaded = 0;
@@ -76,7 +75,7 @@ class FakeImageCacheManager extends Mock implements ImageCacheManager {
   ExpectedData returns(
       String url,
       List<int> imageData, {
-        Duration delayBetweenChunks,
+        Duration? delayBetweenChunks,
       }) {
     const chunkSize = 8;
     final chunks = <Uint8List>[
@@ -86,7 +85,6 @@ class FakeImageCacheManager extends Mock implements ImageCacheManager {
 
     when(getImageFile(
       url,
-      withProgress: anyNamed('withProgress'),
       headers: anyNamed('headers'),
       key: anyNamed('key'),
     )).thenAnswer((realInvocation) => _createResultStream(
@@ -107,7 +105,7 @@ class FakeImageCacheManager extends Mock implements ImageCacheManager {
       String url,
       List<Uint8List> chunks,
       List<int> imageData,
-      Duration delayBetweenChunks,
+      Duration? delayBetweenChunks,
       ) async* {
     var totalSize = imageData.length;
     var downloaded = 0;
@@ -130,5 +128,5 @@ class ExpectedData {
   final int totalSize;
   final int chunkSize;
 
-  const ExpectedData({this.chunks, this.totalSize, this.chunkSize});
+  const ExpectedData({required this.chunks, required this.totalSize, required this.chunkSize});
 }
