@@ -3,28 +3,27 @@ import 'dart:typed_data';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 
 import 'dart:async';
-import 'dart:math' as math;
-import 'dart:ui' show Codec, FrameInfo;
+import 'package:file/memory.dart';
 
 import 'package:cached_network_image/cached_network_image.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/painting.dart';
-import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:flutter_test/flutter_test.dart';
-
+import 'package:mockito/annotations.dart';
+import 'image_cache_manager_test.mocks.dart';
 import 'fake_cache_manager.dart';
 import 'image_data.dart';
 import 'rendering_tester.dart';
 import 'package:mockito/mockito.dart';
 
+@GenerateMocks([CacheManager])
 void main() {
   TestRenderingFlutterBinding();
 
   setUp(() {});
 
   tearDown(() {
-    PaintingBinding.instance.imageCache.clear();
-    PaintingBinding.instance.imageCache.clearLiveImages();
+    PaintingBinding.instance?.imageCache?.clear();
+    PaintingBinding.instance?.imageCache?.clearLiveImages();
   });
 
   test('Supplying an ImageCacheManager should call getImageFile', () async {
@@ -63,8 +62,8 @@ void main() {
   test('Supplying an CacheManager should call getFileStream', () async {
     var url = 'foo.nl';
 
-    var cacheManager = FakeCacheManager();
-    cacheManager.returns(url, kTransparentImage);
+    var cacheManager = MockCacheManager();
+    returns(cacheManager, url, kTransparentImage);
     final imageAvailable = Completer<void>();
 
     final ImageProvider imageProvider =
@@ -90,8 +89,8 @@ void main() {
     var url = 'foo.nl';
     final caughtError = Completer<dynamic>();
 
-    var cacheManager = FakeCacheManager();
-    cacheManager.returns(url, kTransparentImage);
+    var cacheManager = MockCacheManager();
+    returns(cacheManager, url, kTransparentImage);
     final imageAvailable = Completer<void>();
 
     final ImageProvider imageProvider = CachedNetworkImageProvider(url,
@@ -101,7 +100,7 @@ void main() {
     result.addListener(
         ImageStreamListener((ImageInfo image, bool synchronousCall) {
       imageAvailable.complete();
-    }, onError: (dynamic error, StackTrace stackTrace) {
+    }, onError: (dynamic error, StackTrace? stackTrace) {
       caughtError.complete(error);
     }));
     final dynamic err = await caughtError.future;
@@ -113,8 +112,8 @@ void main() {
     var url = 'foo.nl';
     final caughtError = Completer<dynamic>();
 
-    var cacheManager = FakeCacheManager();
-    cacheManager.returns(url, kTransparentImage);
+    var cacheManager = MockCacheManager();
+    returns(cacheManager, url, kTransparentImage);
     final imageAvailable = Completer<void>();
 
     final ImageProvider imageProvider = CachedNetworkImageProvider(url,
@@ -124,11 +123,62 @@ void main() {
     result.addListener(
         ImageStreamListener((ImageInfo image, bool synchronousCall) {
       imageAvailable.complete();
-    }, onError: (dynamic error, StackTrace stackTrace) {
+    }, onError: (dynamic error, StackTrace? stackTrace) {
       caughtError.complete(error);
     }));
     final dynamic err = await caughtError.future;
 
     expect(err, isA<AssertionError>());
   }, skip: isBrowser);
+}
+ExpectedData returns(
+    MockCacheManager cache,
+    String url,
+    List<int> imageData, {
+      Duration? delayBetweenChunks,
+    }) {
+  const chunkSize = 8;
+  final chunks = <Uint8List>[
+    for (int offset = 0; offset < imageData.length; offset += chunkSize)
+      Uint8List.fromList(imageData.skip(offset).take(chunkSize).toList()),
+  ];
+
+  when(cache.getFileStream(
+    url,
+    withProgress: anyNamed('withProgress'),
+    headers: anyNamed('headers'),
+    key: anyNamed('key'),
+  )).thenAnswer((realInvocation) => createResultStream(
+    url,
+    chunks,
+    imageData,
+    delayBetweenChunks,
+  ));
+
+  return ExpectedData(
+    chunks: chunks.length,
+    totalSize: imageData.length,
+    chunkSize: chunkSize,
+  );
+}
+
+Stream<FileResponse> createResultStream(
+    String url,
+    List<Uint8List> chunks,
+    List<int> imageData,
+    Duration? delayBetweenChunks,
+    ) async* {
+  var totalSize = imageData.length;
+  var downloaded = 0;
+  for (var chunk in chunks) {
+    downloaded += chunk.length;
+    if (delayBetweenChunks != null) {
+      await Future.delayed(delayBetweenChunks);
+    }
+    yield DownloadProgress(url, totalSize, downloaded);
+  }
+  var file = MemoryFileSystem().systemTempDirectory.childFile('test.jpg');
+  await file.writeAsBytes(imageData);
+  yield FileInfo(
+      file, FileSource.Online, DateTime.now().add(Duration(days: 1)), url);
 }
